+++ Models/TestKeys.cs
@@ -7,7 +7,7 @@ namespace Console_TelegramBot.Models
         public int ID { set; get; }
         public string Test_Key { set; get; }
         public int Test_ID { get; set; }
-        public int User_ID { set; get; }
+        public long User_ID { set; get; }
         public DateTime Date_Enable_Key { set; get; }
         public DateTime Date_Disable_Key { set; get; }
         public byte Tryes { set; get; }
